examples/requirements.txt
python-pptx>=0.6.21
lxml>=4.9.0
